Batch of CodeQL Fixes
Includes changes across the repo for the following CodeQL rules:

- cpp/comparison-with-wider-type
- cpp/overflow-buffer
- cpp/redundant-null-check-param
- cpp/uselesstest

Some false positives are filtered in CodeQlFilters.yml.

+++ ShellPkg/Library/UefiShellBcfgCommandLib/UefiShellBcfgCommandLib.c
@@ -1020,7 +1020,7 @@ BcfgAddOpt (
   SHELL_STATUS    ShellStatus;
   EFI_STATUS      Status;
   UINT16          OptionIndex;
-  UINT16          LoopCounter;
+  UINT32          LoopCounter;
   UINT64          Intermediate;
   CONST CHAR16    *Temp;
   CONST CHAR16    *Walker;
